+++ Source/Core/VideoCommon/FrameDumpFFMpeg.h
